requirements.txt
absl-py==2.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.33
numpy==2.4.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
sounddevice==0.5.5
pandas
openpyxl
streamlit
pytest
